+++ semgrep_rules_check/eslint_rules/negatives.js
@@ -1,0 +1,5 @@
+function test(someVar, obj) {
+  // detect-non-literal-regexp
+  r1 = new RegExp('boom', 'i')
+  r1 = new RegExp('yeah')
+}