+++ semgrep_rules_check/react_rules/negatives.js
@@ -1,0 +1,3 @@
+React.createElement("div", {}, "Children");
+
+React.createElement("Hello", {}, "Children");