+++ semgrep_rules_check/react_rules/negatives.jsx
@@ -1,0 +1,9 @@
+<div>
+  <div>
+    Children
+  </div>
+
+  <Hello>
+    Children
+  </Hello>
+</div>