+++ semgrep_rules_check/run_semgrep_json.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env sh
+
+semgrep -f ../rules --json . --strict --no-rewrite-rule-ids | python format_json.py